+++ src/services/sheetjs/sheetJSApp.jsx
@@ -8,6 +8,9 @@
 
 import React from "react";
 import XLSX from "xlsx/xlsx";
+import OutTable from "./outTable";
+import DragDropFile from "./dragDropFile";
+import DataInput from "./dataInput";
 
 class SheetJSApp extends React.Component {
     constructor(props) {
@@ -86,137 +89,6 @@ class SheetJSApp extends React.Component {
 
 // if (typeof module !== "undefined") module.exports = SheetJSApp;
 export default SheetJSApp;
-
-/* -------------------------------------------------------------------------- */
-
-/*
-  Simple HTML5 file drag-and-drop wrapper
-  usage: <DragDropFile handleFile={handleFile}>...</DragDropFile>
-    handleFile(file:File):void;
-*/
-class DragDropFile extends React.Component {
-    constructor(props) {
-        super(props);
-        this.onDrop = this.onDrop.bind(this);
-    }
-    suppress(evt) {
-        evt.stopPropagation();
-        evt.preventDefault();
-    }
-    onDrop(evt) {
-        evt.stopPropagation();
-        evt.preventDefault();
-        const files = evt.dataTransfer.files;
-        if (files && files[0]) this.props.handleFile(files[0]);
-    }
-    render() {
-        return (
-            <div
-                onDrop={this.onDrop}
-                onDragEnter={this.suppress}
-                onDragOver={this.suppress}
-            >
-                {this.props.children}
-            </div>
-        );
-    }
-}
-
-/*
-  Simple HTML5 file input wrapper
-  usage: <DataInput handleFile={callback} />
-    handleFile(file:File):void;
-*/
-class DataInput extends React.Component {
-    constructor(props) {
-        super(props);
-        this.handleChange = this.handleChange.bind(this);
-    }
-    handleChange(e) {
-        const files = e.target.files;
-        if (files && files[0]) this.props.handleFile(files[0]);
-    }
-    render() {
-        return (
-            <form className="form-inline">
-                <div className="form-group">
-                    <label htmlFor="file">Spreadsheet</label>
-                    <input
-                        type="file"
-                        className="form-control"
-                        id="file"
-                        accept={SheetJSFT}
-                        onChange={this.handleChange}
-                    />
-                </div>
-            </form>
-        );
-    }
-}
-
-/*
-  Simple HTML Table
-  usage: <OutTable data={data} cols={cols} />
-    data:Array<Array<any> >;
-    cols:Array<{name:string, key:number|string}>;
-*/
-class OutTable extends React.Component {
-    constructor(props) {
-        super(props);
-    }
-    render() {
-        return (
-            <div className="table-responsive">
-                <table className="table table-striped">
-                    <thead>
-                        <tr>
-                            {this.props.cols.map(c => (
-                                <th key={c.key}>{c.name}</th>
-                            ))}
-                        </tr>
-                    </thead>
-                    <tbody>
-                        {this.props.data.map((r, i) => (
-                            <tr key={i}>
-                                {this.props.cols.map(c => (
-                                    <td key={c.key}>{r[c.key]}</td>
-                                ))}
-                            </tr>
-                        ))}
-                    </tbody>
-                </table>
-            </div>
-        );
-    }
-}
-
-/* list of supported file types */
-const SheetJSFT = [
-    "xlsx",
-    "xlsb",
-    "xlsm",
-    "xls",
-    "xml",
-    "csv",
-    "txt",
-    "ods",
-    "fods",
-    "uos",
-    "sylk",
-    "dif",
-    "dbf",
-    "prn",
-    "qpw",
-    "123",
-    "wb*",
-    "wq*",
-    "html",
-    "htm"
-]
-    .map(function(x) {
-        return "." + x;
-    })
-    .join(",");
 
 /* generate an array of column objects */
 const make_cols = refstr => {
